Finished Assignment
Fixed the methods that added to the doubly linked list

added pthread mutex to stop threads for changing the list at the same time

added pthread cond to wait when buffer is full, buffer is empty, or when the consumer is waiting for an odd or even number

Added the ability to signal once the buffer gets changed

join the threads after creating so that main thread has to wait for these new threads to finish. stops premature ending.